fix: rm

+++ include/io.h
@@ -4,7 +4,7 @@
 #define ENDERECO_SIZE (uint8_t)8
 #define ATRIBUTOS_SIZE (uint8_t)1
 #define NOME_SIZE (uint16_t)256
-#define ARQUIVO_SIZE (FLAGS_SIZE + ENDERECO_SIZE + ATRIBUTOS_SIZE + NOME_SIZE)
+#define ARQUIVO_SIZE (uint16_t)(FLAGS_SIZE + ENDERECO_SIZE + ATRIBUTOS_SIZE + NOME_SIZE)
 
 #include "boolean.h"
 #include <stdint.h>
